refactor: 재사용되는 ArtworkListParams 타입을 분리하여 따로 관리
